xavier_mvp_booster_pack_v0.1.0/requirements.txt
jsonschema>=4.22.0
pynacl>=1.5.0
streamlit>=1.36.0
matplotlib>=3.8.0
numpy>=1.26.0
